Split the web example into modules, mention mudularity into docs
