+++ examples/web/pinger/services.go
@@ -1,0 +1,14 @@
+package pinger
+
+import (
+	"github.com/zhulik/pal"
+	"github.com/zhulik/pal/examples/web/core"
+)
+
+// Provide provides the pinger service.
+func Provide() pal.ServiceDef {
+	return pal.ProvideList(
+		pal.Provide[core.Pinger](&Pinger{}),
+		// Add more services here, if needed.
+	)
+}
