+++ examples/web/server/services.go
@@ -1,0 +1,13 @@
+package server
+
+import (
+	"github.com/zhulik/pal"
+)
+
+// Provide provides the server service.
+func Provide() pal.ServiceDef {
+	return pal.ProvideList(
+		pal.Provide(&Server{}),
+		// Add more services here, if needed.
+	)
+}
